coreapi: move path impl to path.go
License: MIT
